blood implemented

+++ Warcraft Heroes Beyond Time/Warcraft Heroes Beyond Time/GUIImage.h
@@ -18,7 +18,7 @@ public:
 	}
 
 private:
-	uint opacity = 255;
+	Uint8 opacity = 255;
 };
 
 #endif
